+++ mllib/src/main/scala/org/apache/spark/ml/image/ImageSchema.scala
@@ -191,76 +191,4 @@ object ImageSchema {
       Some(Row(Row(origin, height, width, nChannels, mode, decoded)))
     }
   }
-
-  /**
-   * Read the directory of images from the local or remote source
-   *
-   * @note If multiple jobs are run in parallel with different sampleRatio or recursive flag,
-   * there may be a race condition where one job overwrites the hadoop configs of another.
-   * @note If sample ratio is less than 1, sampling uses a PathFilter that is efficient but
-   * potentially non-deterministic.
-   *
-   * @param path Path to the image directory
-   * @return DataFrame with a single column "image" of images;
-   *         see ImageSchema for the details
-   */
-  @deprecated("use `spark.read.format(\"image\").load(path)` and this `readImages` will be " +
-    "removed in 3.0.0.", "2.4.0")
-  def readImages(path: String): DataFrame = readImages(path, null, false, -1, false, 1.0, 0)
-
-  /**
-   * Read the directory of images from the local or remote source
-   *
-   * @note If multiple jobs are run in parallel with different sampleRatio or recursive flag,
-   * there may be a race condition where one job overwrites the hadoop configs of another.
-   * @note If sample ratio is less than 1, sampling uses a PathFilter that is efficient but
-   * potentially non-deterministic.
-   *
-   * @param path Path to the image directory
-   * @param sparkSession Spark Session, if omitted gets or creates the session
-   * @param recursive Recursive path search flag
-   * @param numPartitions Number of the DataFrame partitions,
-   *                      if omitted uses defaultParallelism instead
-   * @param dropImageFailures Drop the files that are not valid images from the result
-   * @param sampleRatio Fraction of the files loaded
-   * @return DataFrame with a single column "image" of images;
-   *         see ImageSchema for the details
-   */
-  @deprecated("use `spark.read.format(\"image\").load(path)` and this `readImages` will be " +
-    "removed in 3.0.0.", "2.4.0")
-  def readImages(
-      path: String,
-      sparkSession: SparkSession,
-      recursive: Boolean,
-      numPartitions: Int,
-      dropImageFailures: Boolean,
-      sampleRatio: Double,
-      seed: Long): DataFrame = {
-    require(sampleRatio <= 1.0 && sampleRatio >= 0, "sampleRatio should be between 0 and 1")
-
-    val session = if (sparkSession != null) sparkSession else SparkSession.builder().getOrCreate
-    val partitions =
-      if (numPartitions > 0) {
-        numPartitions
-      } else {
-        session.sparkContext.defaultParallelism
-      }
-
-    RecursiveFlag.withRecursiveFlag(recursive, session) {
-      SamplePathFilter.withPathFilter(sampleRatio, session, seed) {
-        val binResult = session.sparkContext.binaryFiles(path, partitions)
-        val streams = if (numPartitions == -1) binResult else binResult.repartition(partitions)
-        val convert = (origin: String, bytes: PortableDataStream) =>
-          decode(origin, bytes.toArray())
-        val images = if (dropImageFailures) {
-          streams.flatMap { case (origin, bytes) => convert(origin, bytes) }
-        } else {
-          streams.map { case (origin, bytes) =>
-            convert(origin, bytes).getOrElse(invalidImageRow(origin))
-          }
-        }
-        session.createDataFrame(images, imageSchema)
-      }
-    }
-  }
 }
