+++ python/pyspark/ml/tests/test_image.py
@@ -24,18 +24,24 @@ from pyspark.sql import HiveContext, Row
 from pyspark.testing.utils import QuietTest
 
 
-class ImageReaderTest(SparkSessionTestCase):
+class ImageFileFormatTest(SparkSessionTestCase):
 
     def test_read_images(self):
         data_path = 'data/mllib/images/origin/kittens'
-        df = ImageSchema.readImages(data_path, recursive=True, dropImageFailures=True)
+        df = self.spark.read.format("image") \
+            .option("dropInvalid", True) \
+            .option("recursiveFileLookup", True) \
+            .load(data_path)
         self.assertEqual(df.count(), 4)
         first_row = df.take(1)[0][0]
+        # compare `schema.simpleString()` instead of directly compare schema,
+        # because the df loaded from datasouce may change schema column nullability.
+        self.assertEqual(df.schema.simpleString(), ImageSchema.imageSchema.simpleString())
+        self.assertEqual(df.schema["image"].dataType.simpleString(),
+                         ImageSchema.columnSchema.simpleString())
         array = ImageSchema.toNDArray(first_row)
         self.assertEqual(len(array), first_row[1])
         self.assertEqual(ImageSchema.toImage(array, origin=first_row[0]), first_row)
-        self.assertEqual(df.schema, ImageSchema.imageSchema)
-        self.assertEqual(df.schema["image"].dataType, ImageSchema.columnSchema)
         expected = {'CV_8UC3': 16, 'Undefined': -1, 'CV_8U': 0, 'CV_8UC1': 0, 'CV_8UC4': 24}
         self.assertEqual(ImageSchema.ocvTypes, expected)
         expected = ['origin', 'height', 'width', 'nChannels', 'mode', 'data']
@@ -61,11 +67,11 @@ class ImageReaderTest(SparkSessionTestCase):
                 lambda: ImageSchema.toImage("a"))
 
 
-class ImageReaderTest2(PySparkTestCase):
+class ImageFileFormatOnHiveContextTest(PySparkTestCase):
 
     @classmethod
     def setUpClass(cls):
-        super(ImageReaderTest2, cls).setUpClass()
+        super(ImageFileFormatOnHiveContextTest, cls).setUpClass()
         cls.hive_available = True
         # Note that here we enable Hive's support.
         cls.spark = None
@@ -86,17 +92,20 @@ class ImageReaderTest2(PySparkTestCase):
 
     @classmethod
     def tearDownClass(cls):
-        super(ImageReaderTest2, cls).tearDownClass()
+        super(ImageFileFormatOnHiveContextTest, cls).tearDownClass()
         if cls.spark is not None:
             cls.spark.sparkSession.stop()
             cls.spark = None
 
     def test_read_images_multiple_times(self):
-        # This test case is to check if `ImageSchema.readImages` tries to
+        # This test case is to check if ImageFileFormat tries to
         # initiate Hive client multiple times. See SPARK-22651.
         data_path = 'data/mllib/images/origin/kittens'
-        ImageSchema.readImages(data_path, recursive=True, dropImageFailures=True)
-        ImageSchema.readImages(data_path, recursive=True, dropImageFailures=True)
+        for i in range(2):
+            self.spark.read.format("image") \
+                .option("dropInvalid", True) \
+                .option("recursiveFileLookup", True) \
+                .load(data_path)
 
 
 if __name__ == "__main__":
